[LCR-349] refactor: page index.tsx 로 변경
